Add ENABLE_CRASH_HANDLER compile option
The current dwarfstack-based crash handler is Windows-only. This moves it to an
isolated source file.
