+++ Makefile
@@ -7,7 +7,7 @@ CC=gcc
 OPT=-O3
 DEBUG=-gdwarf-4
 SDL_CFLAGS=$(shell $(SDL_CONFIG) --cflags)
-CFLAGS=$(OPT) $(DEBUG) -Wall -Wno-pointer-sign -Wno-char-subscripts -fno-omit-frame-pointer -fvisibility=hidden $(SDL_CFLAGS)
+CFLAGS=$(OPT) $(DEBUG) -Wall -Wno-pointer-sign -Wno-char-subscripts -fno-omit-frame-pointer -fvisibility=hidden -DENABLE_CRASH_HANDLER $(SDL_CFLAGS)
 LDFLAGS=$(OPT) $(DEBUG) -Wl,-subsystem,windows
 
 SDL_LIBS=$(shell $(SDL_CONFIG) --libs)
@@ -18,7 +18,7 @@ OBJS	=		src/gui/gui.o src/client/client.o src/client/skill.o src/game/dd.o src/g
 			src/sdl/sound.o src/game/resource.o src/sdl/sdl.o src/helper/helper.o\
 			src/gui/dots.o src/gui/display.o src/gui/teleport.o src/gui/color.o src/gui/cmd.o\
 			src/gui/questlog.o src/gui/context.o src/gui/hover.o src/modder/sharedmem.o\
-			src/gui/minimap.o
+			src/gui/minimap.o src/game/crash_handler_windows.o
 
 bin/moac.exe lib/moac.a &:	$(OBJS)
 			$(CC) $(LDFLAGS) -Wl,--out-implib,lib/moac.a -o bin/moac.exe $(OBJS) src/game/version.c $(LIBS)
@@ -63,6 +63,8 @@ src/modder/sharedmem.o:	src/modder/sharedmem.c src/astonia.h src/modder.h src/mo
 
 src/sdl/sdl.o:		src/sdl/sdl.c src/astonia.h src/sdl.h src/sdl/_sdl.h
 src/sdl/sound.o:      	src/sdl/sound.c src/astonia.h src/sdl.h src/sdl/_sdl.h
+
+src/game/crash_handler_windows.o: src/game/crash_handler_windows.c
 
 src/game/resource.o:	src/game/resource.rc src/game/resource.h res/moa3.ico
 			$(WINDRES) -F pe-x86-64 src/game/resource.rc src/game/resource.o
